WIP: Add wallclock time profiling.
With this, if you send SIGUSR1 to the pageserver, it will spit out a
flamegraph at '/tmp/flamegraph.svg'. The flamegraph includes time spent
on processing and waiting on I/O or locks, but not the time spent waiting
for request from client.

By commenting/uncommenting the 'pprof_guard.start();' calls, you can
enable profiling for get_page_at_lsn() requests, or for WAL receiver or
for checkpoints.

This depends on modified version of 'pprof-rs' crate from here:
https://github.com/hlinnaka/pprof-rs/tree/wallclock-profiling-wip

This also includes a bunch of small code changes that I was experimenting
with that are unrelated to profiling. Sorry about the mess.